LaplaceInterpolation.jl Documentation

Complete documentation is under construction, for now, here is an index of the function signatures.

Punching

LaplaceInterpolation.punch_holes_3D — Function

punchholes3D(centers, radius, xpoints, ypoints, zpoints)

...

Arguments

centers::Vector{T}: the vector containing the centers of the punches

radius::Float64: the radius of the punch

Nx::Int64: the number of points in the x-direction, this code is hard-coded to start from one.

Ny::Int64: the number of points in the y-direction

Nz::Int64: the number of points in the z-direction

...

...

Outputs

absolute_indices::Vector{Int64}: vector containing the indices of coordinates

inside the punch ...

LaplaceInterpolation.punch_holes_2D — Function

punchholes2D(centers, radius, xpoints, ypoints)

...

Arguments

centers::Union{Vector}: the vector containing the punch centers

radius::Vector: the tuple containing the punch radii

Nx::Int64: the number of points in the x direction

Ny::Int64: the number of points in the y direction

...

...

Outputs

absolute_indices::Vector{Int64}: vector containing the indices of coordinates

inside the punch ...

Missing docstring.

Missing docstring for punch_holes_nexus. Check Documenter's build log for details.

One dimensional

LaplaceInterpolation.nablasq_grid — Function

nablasq_grid(n, h)

Laplacian matrix on a 1D grid

Arguments:

- `n`: Number of points
- `h`: Aspect ratio

Outputs:

- discrete Laplacian matrix

nablasq_grid(Nx, Ny, h, k)

Laplacian matrix on a 2D grid

Arguments:

- `Nx::Int64`: Number of points in first dimension
- `Ny::Int64`: Number of points in second dimension
- `h::Float64`: Aspect ratio in first dimension
- `k::Float64`: Aspect ratio in second dimension

Outputs:

- discrete Laplacian matrix in 2D

Missing docstring.

Missing docstring for matern_1D_grid. Check Documenter's build log for details.

Two dimensional

LaplaceInterpolation.bdy_nodes — Function

bdy_nodes(Nx, Ny)

...

Arguments

Nx::Int64: the number of points in the first dimension

Ny::Int64: the number of points in the second dimension

...

...

Outputs

vector containing the indices of coordinates on the boundary of the 2D rectangle

...

bdy_nodes(dims)

Boundary node computation, for arbitrary dimension

...

Arguments

dims::Tuple number of points in each direction

...

...

Outputs

Vector{Int64}: vector containing the indices of coordinates

on the boundary of the hyperrectangle volume ...

LaplaceInterpolation.matern_2d_grid — Function

matern2dgrid(mat, discard, m, eps, h, k)

...

Arguments

mat: the matrix containing the image

idx: the linear indices of the nodes to be discarded

eps: Matern parameter eps

m: The Matern exponent (integer)

h: The aspect ratio in the first dimension

k: The aspect ratio in the second dimension

Outputs

matrix containing the interpolated image

Example:

```<julia-repl> x = y = 1:30 h = k = x[2] - x[1] y = sin.(2 * pi * x * 0.2) * cos.(2 * pi * y * 0.3) discard = randperm(900)[1:450]

Laplace interpolation

ylap = matern2d_grid(y, discard, 1, 0.0, h, k)

Matern interpolation

ymat = matern2d_grid(y, discard, 2, 0.1, h, k) ...

Three diimensional

LaplaceInterpolation.nablasq_3d_grid — Function

nablasq3dgrid(Nx,Ny)

Construct the 3D Laplace matrix

Arguments

Nx::Int64: The number of nodes in the first dimension

Ny::Int64: The number of nodes in the second dimension

Nz::Int64: The number of nodes in the third dimension

h::Float64: Grid spacing in the first dimension

k::Float64: Grid spacing in the second dimension

l::Float64: Grid spacing in the third dimension

Outputs

-nablasq (discrete Laplacian, real-symmetric positive-definite) on Nx×Ny×Nz grid

LaplaceInterpolation.matern_3d_grid — Function

matern3dgrid(imgg, discard, m, eps, h, k, l)

Interpolates a single punch

...

Arguments

imgg: the matrix containing the image

discard::Union{Vector{CartesianIndex{3}}}, Vector{Int64}}: the linear or Cartesian indices of the values to be filled

m::Int64 = 1 : Matern parameter

eps::Float64 = 0.0: Matern parameter eps

h = 1.0: Aspect ratio in the first dimension

k = 1.0: Aspect ratio in the second dimension

l = 1.0: Aspect ratio in the third dimension

Outputs

array containing the restored image

...

Missing docstring.

Missing docstring for Laplace_3D_Grid. Check Documenter's build log for details.

Missing docstring.

Missing docstring for parallel_Matern_3DGrid. Check Documenter's build log for details.

Missing docstring.

Missing docstring for parallel_Laplace_3Dgrid. Check Documenter's build log for details.

Missing docstring.

Missing docstring for interp_nexus. Check Documenter's build log for details.

LaplaceInterpolation.spdiagm_nonsquare — Function

spdiagm_nonsquare(m, n, args...)

Construct a sparse diagonal matrix from Pairs of vectors and diagonals. Each vector arg.second will be placed on the arg.first diagonal. By default (if size=nothing), the matrix is square and its size is inferred from kv, but a non-square size m×n (padded with zeros as needed) can be specified by passing m,n as the first arguments.

Arguments

m::Int64: First dimension of the output matrix

n::Int64: Second dimension of the output matrix

args::Tuple{T} where T<:Pair{<:Integer,<:AbstractVector}

Outputs

sparse matrix of size mxn containing the values in args

LaplaceInterpolation.return_boundary_nodes — Function

returnboundarynodes(xpoints, ypoints, zpoints)

...

Arguments

xpoints::Vector{T} where T<:Real: the vector containing the x coordinate

ypoints::Vector{T} where T<:Real: the vector containing the y coordinate

zpoints::Vector{T} where T<:Real: the vector containing the z coordinate

...

...

Outputs

BoundaryNodes3D::Vector{Int64}: vector containing the indices of coordinates

on the boundary of the rectangular 3D volume ...

LaplaceInterpolation.Matern3D_Grid — Function

Matern3D_Grid(xpoints, ypoints, zpoints, imgg, epsilon, radius, h, k, l, m)

...

Arguments

xpoints::Vector{T} where T<:Real: the vector containing the x coordinate

ypoints::Vector{T} where T<:Real: the vector containing the y coordinate

zpoints::Vector{T} where T<:Real: the vector containing the z coordinate

imgg: the matrix containing the image

epsilon: Matern parameter epsilon

radius::Vector: the tuple containing the punch radii

h::Float: grid spacing along the x-axis

k::Float: grid spacing along the y-axis

l::Float: grid spacing along the z-axis

m::Int : Matern parameter

Outputs

tuple containing the restored image and the punched image.

...

LaplaceInterpolation.Parallel_Matern3D_Grid — Function

ParallelMatern3DGrid(xpoints, ypoints, zpoints, imgg, epsilon, radius, h, k, l, xmin, xmax, ymin, ymax, zmin, zmax, m)

Compute the spherically-punched, Matern-interpolated 3D data

...

Arguments

xpoints::Vector{T} where T<:Real: the vector containing the x coordinate

ypoints::Vector{T} where T<:Real: the vector containing the y coordinate

zpoints::Vector{T} where T<:Real: the vector containing the z coordinate

imgg: the matrix containing the image

epsilon: one of the matern parameters

radius::Vector: the tuple containing the punch radii

h::Float: grid spacing along the x-axis

k::Float: grid spacing along the y-axis

l::Float: grid spacing along the z-axis

xmin::Int64: Vishwas should fill in the next six fields.

xmax::Int64:

ymin::Int64:

ymax::Int64:

zmin::Int64:

zmax::Int64:

m::Int64: The matern order parameter

Outputs

array containing the restored image

Example

(xmin, xmax) = (ymin, ymax) = (zmin,zmax) = (0.0, 1.0)
xpoints = ypoints = zpoints = -0.2:0.2:1.2
h = k = l = 0.2
imgg = randn(8,8,8)
m = 2
epsilon = 0.0
radius = 0.2
restored = Parallel_Matern3D_Grid(xpoints, ypoints, zpoints, imgg, epsilon, radius, 
                                  h, k, l, xmin, xmax, ymin, ymax, zmin, zmax, m)

...